It simulates with valid output
Although the output it not exactly correct, they are no longer all 'x so
that is a good start. At this point I can verify that loading the
instruction from the pc value into the instruction reg works. I'm not
sure that the incrementing of the pc is working properly but I will
verify that tomorrow and then on to testplan and test cases